Renamed IndexController to HomeController
